Use NetworkStream instead of StreamReaders/Writers
Client:
- refac: For reading and writing, use the NetworkStream instead of
StreamReaders/Writers to read and write data. This makes it possible to
read newline characters since Read() reads every byte instead of a
single line. The same goes for Write().
-fix: Possible data race where the read data was overwritten during the
next loop iteration and the data was not yet passed to Task.Run(). Now,
a check is performed to see if there is data available. Only if there is
data to read, a buffer is created which cannot be set to null the next
iteration.